+++ crates/rust-client/src/store/mod.rs
@@ -28,7 +28,7 @@ use alloc::{
 use core::fmt::Debug;
 
 use miden_objects::{
-    Digest, Word,
+    Digest, MastForest, Word,
     account::{Account, AccountCode, AccountHeader, AccountId},
     block::{BlockHeader, BlockNumber},
     crypto::merkle::{InOrderIndex, MmrPeaks},
@@ -276,6 +276,13 @@ pub trait Store: Send + Sync {
     ///
     /// Returns a `StoreError::AccountDataNotFound` if there is no account for the provided ID.
     async fn update_account(&self, new_account_state: &Account) -> Result<(), StoreError>;
+
+    /// Retrieves the MAST forest that contains the provided procedure root. If the forest is not
+    /// found, returns `None`.
+    async fn get_mast_forest(
+        &self,
+        procedure_root: Digest,
+    ) -> Result<Option<MastForest>, StoreError>;
 
     // SYNC
     // --------------------------------------------------------------------------------------------