+++ crates/rust-client/src/store/web_store/account/js_bindings.rs
@@ -39,7 +39,12 @@ extern "C" {
     // ================================================================================================
 
     #[wasm_bindgen(js_name = insertAccountCode)]
-    pub fn idxdb_insert_account_code(code_root: String, code: Vec<u8>) -> js_sys::Promise;
+    pub fn idxdb_insert_account_code(
+        code_root: String,
+        mast: Vec<u8>,
+        procedure_info: Vec<u8>,
+        procedure_roots: Vec<String>,
+    ) -> js_sys::Promise;
 
     #[wasm_bindgen(js_name = insertAccountStorage)]
     pub fn idxdb_insert_account_storage(
@@ -69,12 +74,20 @@ extern "C" {
     #[wasm_bindgen(js_name = upsertForeignAccountCode)]
     pub fn idxdb_upsert_foreign_account_code(
         account_id: String,
-        code: Vec<u8>,
         code_root: String,
+        mast: Vec<u8>,
+        procedure_info: Vec<u8>,
+        procedure_roots: Vec<String>,
     ) -> js_sys::Promise;
 
     #[wasm_bindgen(js_name = getForeignAccountCode)]
     pub fn idxdb_get_foreign_account_code(account_ids: Vec<String>) -> js_sys::Promise;
+
+    #[wasm_bindgen(js_name = getMastForest)]
+    pub fn idxdb_get_mast_forest(procedure_root: String) -> JsValue;
+
+    #[wasm_bindgen(js_name = fetchAndCacheMastForests)]
+    pub fn idxdb_fetch_and_cache_mast_forests() -> js_sys::Promise;
 
     // UPDATES
     // ================================================================================================