+++ crates/rust-client/src/store/web_store/account/mod.rs
@@ -1,12 +1,15 @@
 use alloc::{
     collections::BTreeMap,
     string::{String, ToString},
+    sync::Arc,
     vec::Vec,
 };
 
 use miden_objects::{
-    AccountIdError, Digest, Word,
-    account::{Account, AccountCode, AccountHeader, AccountId, AccountStorage},
+    AccountIdError, Digest, MastForest, Word,
+    account::{
+        Account, AccountCode, AccountHeader, AccountId, AccountProcedureInfo, AccountStorage,
+    },
     asset::{Asset, AssetVault},
 };
 use miden_tx::utils::{Deserializable, Serializable};
@@ -14,7 +17,13 @@ use serde_wasm_bindgen::from_value;
 use wasm_bindgen_futures::JsFuture;
 
 use super::WebStore;
-use crate::store::{AccountRecord, AccountStatus, StoreError};
+use crate::store::{
+    AccountRecord, AccountStatus, StoreError,
+    web_store::account::{
+        js_bindings::{idxdb_fetch_and_cache_mast_forests, idxdb_get_mast_forest},
+        models::MastForestIdxdbObject,
+    },
+};
 
 mod js_bindings;
 use js_bindings::{
@@ -160,8 +169,10 @@ impl WebStore {
         let account_code_idxdb: AccountCodeIdxdbObject = from_value(js_value)
             .map_err(|err| StoreError::DatabaseError(format!("failed to deserialize {err:?}")))?;
 
-        let code =
-            AccountCode::from_bytes(&account_code_idxdb.code).map_err(StoreError::AccountError)?;
+        let mast = MastForest::read_from_bytes(&account_code_idxdb.mast)?;
+        let procedure_info =
+            Vec::<AccountProcedureInfo>::read_from_bytes(&account_code_idxdb.procedure_info)?;
+        let code = AccountCode::from_parts(Arc::new(mast), procedure_info);
 
         Ok(code)
     }
@@ -263,16 +274,40 @@ impl WebStore {
         }
     }
 
+    pub async fn fetch_and_cache_mast_forests(&self) -> Result<(), StoreError> {
+        let promise = idxdb_fetch_and_cache_mast_forests();
+        JsFuture::from(promise).await.map_err(|js_error| {
+            StoreError::DatabaseError(format!(
+                "failed to fetch and cache mast forests: {js_error:?}",
+            ))
+        })?;
+
+        Ok(())
+    }
+
     pub(crate) async fn upsert_foreign_account_code(
         &self,
         account_id: AccountId,
         code: AccountCode,
     ) -> Result<(), StoreError> {
-        let root = code.commitment().to_string();
-        let code = code.to_bytes();
         let account_id = account_id.to_string();
+        let root = code.commitment().to_string();
+        let mast_forest = code.mast().to_bytes();
+        let procedure_info = code.procedures().to_vec().to_bytes();
+        let procedure_roots = code
+            .procedures()
+            .iter()
+            .map(|proc| proc.mast_root().to_hex())
+            .collect::<Vec<String>>();
 
-        let promise = idxdb_upsert_foreign_account_code(account_id, code, root);
+        let promise = idxdb_upsert_foreign_account_code(
+            account_id,
+            root,
+            mast_forest,
+            procedure_info,
+            procedure_roots,
+        );
+
         JsFuture::from(promise).await.map_err(|js_error| {
             StoreError::DatabaseError(format!(
                 "failed to upsert foreign account code: {js_error:?}",
@@ -306,14 +341,36 @@ impl WebStore {
             .map(|idxdb_object| {
                 let account_id = AccountId::from_hex(&idxdb_object.account_id)
                     .map_err(StoreError::AccountIdError)?;
-                let code = AccountCode::from_bytes(&idxdb_object.code)
-                    .map_err(StoreError::AccountError)?;
+
+                let mast = MastForest::read_from_bytes(&idxdb_object.mast)?;
+                let procedure_info =
+                    Vec::<AccountProcedureInfo>::read_from_bytes(&idxdb_object.procedure_info)?;
+                let code = AccountCode::from_parts(Arc::new(mast), procedure_info);
 
                 Ok((account_id, code))
             })
             .collect::<Result<BTreeMap<AccountId, AccountCode>, StoreError>>()?;
 
         Ok(foreign_account_code)
+    }
+
+    #[allow(clippy::unused_self)]
+    pub(crate) fn get_mast_forest(
+        &self,
+        procedure_root: Digest,
+    ) -> Result<Option<MastForest>, StoreError> {
+        let procedure_root_str = procedure_root.to_string();
+        let js_value = idxdb_get_mast_forest(procedure_root_str);
+
+        if js_value.is_null() || js_value.is_undefined() {
+            return Ok(None);
+        }
+
+        let mast_forest_idxdb: MastForestIdxdbObject = from_value(js_value)
+            .map_err(|err| StoreError::DatabaseError(format!("failed to deserialize {err:?}")))?;
+        let mast_forest = MastForest::read_from_bytes(&mast_forest_idxdb.mast)?;
+
+        Ok(Some(mast_forest))
     }
 
     pub(crate) async fn undo_account_states(